Begin Testing
So far initial tests have been relatively successful.  The clients are
able to connect to the server and send messages to each other. Minor
editing within the displayed text. Some errors: A repetitive text is
displayed on the client views if the server is closed out of during
connection. Also the server still needs to display client messages.